Not call function that causes crash
This is a very dumb fix, but it works while my understanding of the game
code is still low. The game before assigining all of the parts of the
library checks if that library exists already in memory, if it does it
will call a class function that seems to pertain to the objects
properties.

If the modloader was active this function would fail because the this
pointer would be garbage when calling it. To circunvent this I simply
don't call the function anymore. A better fix would be to investigate
why the this pointer is malformed or only call the function when it's
the expected value (I would assume the global library would always be
the correct one?)

This fixes #1
Further investigation is needed to determine if this affects #2 too, as
I have a suspicion that it might.